+++ src/main/java/com/leccion_3/repositorio/CategoriasRepositorio.java
@@ -6,6 +6,7 @@
 package com.leccion_3.repositorio;
 
 import com.leccion_3.model.Categoria;
+import java.util.Optional;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -14,7 +15,9 @@ import org.springframework.stereotype.Repository;
  * @author igorr
  */
 @Repository
-public interface CategoriasRepositorio extends CrudRepository<Categoria, Long>{
+public interface CategoriasRepositorio extends CrudRepository<Categoria, Integer>{
+
+    
     //el repositorio crudrepositori ya tiene una serie esenciales de metodos implementados , 
     //para trabajar con bbdd pero si se necesita un metodo especifico se implementa aqui mismito!!
     //para ver metodos pinchar con ctrl + click izquierdo
